continue breaking up big main
